# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clinics/me:
    get:
      tags: [Clinics]
      operationId: getMyClinic
      summary: Get the authenticated user's clinic, if it exists
      responses:
        '200':
          description: Clinic returned (or null if not yet set up).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClinicEnvelope'
    patch:
      tags: [Clinics]
      operationId: updateMyClinic
      summary: Update the authenticated user's clinic settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateClinicRequest'
      responses:
        '200':
          description: Updated clinic.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Clinic'
  /clinics:
    post:
      tags: [Clinics]
      operationId: createMyClinic
      summary: Create the clinic for the authenticated user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClinicRequest'
      responses:
        '201':
          description: Clinic created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Clinic'
...
  /patients:
    get:
      tags: [Patients]
      operationId: listMyQueue
      summary: List active patients in the authenticated clinic's queue
      responses:
        '200':
          description: Patients waiting or currently being served.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Patient'
    post:
      tags: [Patients]
      operationId: addPatientToQueue
      summary: Add a patient to the queue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddPatientRequest'
      responses:
        '201':
          description: Patient added.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
...
  /patients/{id}:
    delete:
      tags: [Patients]
      operationId: removePatient
      summary: Remove a patient from the queue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Removed.
  /patients/{id}/skip:
    post:
      tags: [Patients]
      operationId: skipPatient
      summary: Mark a patient as skipped
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Skipped.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
  /public/clinics/{slug}:
    get:
      tags: [Public]
      operationId: getPublicClinic
      summary: Get public-facing info for a clinic
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Public clinic info.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicClinic'
        '404':
          description: Clinic not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /public/clinics/{slug}/join:
    post:
      tags: [Public]
      operationId: publicJoinQueue
      summary: Public self-registration into a clinic's queue (used after WhatsApp scan)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublicJoinRequest'
      responses:
        '201':
          description: Patient added; returns confirmation payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicJoinResponse'
        '404':
          description: Clinic not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /public/queue/{trackingCode}:
    get:
      tags: [Public]
      operationId: getPublicTracking
      summary: Get live tracking info for a patient
      parameters:
        - name: trackingCode
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tracking info.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicTracking'
        '404':
          description: Patient not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Clinic:
      type: object
      required:
        - id
        - slug
        - name
        - doctorName
        - avgConsultationMinutes
        - whatsappNumber
        - createdAt
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        doctorName:
          type: string
        avgConsultationMinutes:
          type: integer
        whatsappNumber:
          type: string
          description: E.164 style phone number, no plus sign required
        createdAt:
          type: string
          format: date-time
    ClinicEnvelope:
      type: object
      required: [clinic]
      properties:
        clinic:
          oneOf:
            - $ref: '#/components/schemas/Clinic'
            - type: 'null'
    CreateClinicRequest:
      type: object
      required: [name, doctorName, avgConsultationMinutes, whatsappNumber]
      properties:
        name:
          type: string
          minLength: 1
        doctorName:
          type: string
          minLength: 1
        avgConsultationMinutes:
          type: integer
          minimum: 1
          maximum: 120
        whatsappNumber:
          type: string
          minLength: 6
    UpdateClinicRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        doctorName:
          type: string
          minLength: 1
        avgConsultationMinutes:
          type: integer
          minimum: 1
          maximum: 120
        whatsappNumber:
          type: string
          minLength: 6
...
    Patient:
      type: object
      required:
        - id
        - clinicId
        - name
        - phone
        - tokenNumber
        - status
        - trackingCode
        - position
        - estimatedWaitMinutes
        - reminderStage
        - createdAt
      properties:
        id:
          type: string
        clinicId:
          type: string
        name:
          type: string
        phone:
          type: string
        tokenNumber:
          type: integer
        status:
          type: string
          enum: [waiting, in_progress, done, skipped]
        trackingCode:
          type: string
        position:
          type: integer
          description: 0 = currently being served, 1 = next, etc.
        estimatedWaitMinutes:
          type: integer
        reminderStage:
          type: string
          enum: [none, three_away, two_away, one_away, your_turn, done]
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: [string, 'null']
          format: date-time
    AddPatientRequest:
      type: object
      required: [name, phone]
      properties:
        name:
          type: string
          minLength: 1
        phone:
          type: string
          minLength: 4
...
    PublicClinic:
      type: object
      required:
        - id
        - slug
        - name
        - doctorName
        - whatsappNumber
        - avgConsultationMinutes
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        doctorName:
          type: string
        whatsappNumber:
          type: string
        avgConsultationMinutes:
          type: integer
    PublicJoinRequest:
      type: object
      required: [name, phone]
      properties:
        name:
          type: string
          minLength: 1
        phone:
          type: string
          minLength: 4
    PublicJoinResponse:
      type: object
      required:
        - clinicName
        - doctorName
        - tokenNumber
        - estimatedWaitMinutes
        - position
        - trackingCode
        - trackingUrl
        - whatsappConfirmationMessage
      properties:
        clinicName:
          type: string
        doctorName:
          type: string
        tokenNumber:
          type: integer
        estimatedWaitMinutes:
          type: integer
        position:
          type: integer
        trackingCode:
          type: string
        trackingUrl:
          type: string
        whatsappConfirmationMessage:
          type: string
    PublicTracking:
      type: object
      required:
        - clinicName
        - doctorName
        - tokenNumber
        - status
        - position
        - patientsAhead
        - estimatedWaitMinutes
        - currentTokenNumber
        - reminderStage
        - avgConsultationMinutes
      properties:
        clinicName:
          type: string
        doctorName:
          type: string
        tokenNumber:
          type: integer
        status:
          type: string
          enum: [waiting, in_progress, done, skipped]
        position:
          type: integer
        patientsAhead:
          type: integer
        estimatedWaitMinutes:
          type: integer
        currentTokenNumber:
          type: [integer, 'null']
        reminderStage:
          type: string
          enum: [none, three_away, two_away, one_away, your_turn, done]
        avgConsultationMinutes:
          type: integer